requirements.txt
beautifulsoup4==4.12.2
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
Flask==2.3.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
graphql-core==3.2.3
gunicorn==21.2.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.3
MarkupSafe==2.1.3
numpy==1.26.0
packaging==23.2
pandas==2.1.1
pika==1.3.2
protobuf==4.22.3
psycopg2-binary==2.9.7
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
sgqlc==16.3
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.21
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.5
Werkzeug==2.3.7
